+++ tests/shell/008-only.sh
@@ -12,9 +12,9 @@ echo 'packaging>=20' > requirements.txt
 echo '{}' > package.json
 
 # Only the given file names are considered:
-upd --dry-run --only requirements.txt | grep -q "requirements.txt"
-! upd --dry-run --only requirements.txt | grep -q "package.json"
+updee --dry-run --only requirements.txt | grep -q "requirements.txt"
+! updee --dry-run --only requirements.txt | grep -q "package.json"
 
 # Unknown file names are an error:
-! upd --dry-run --only foo.txt
-upd --dry-run --only foo.txt | grep -q "unknown file name"
+! updee --dry-run --only foo.txt
+updee --dry-run --only foo.txt | grep -q "unknown file name"